Fix UnicodeEncodeError
(Might have been a UnicodeDecodeError--I forget which it was.)

Sometimes HTMLParser parses out a unicode for reasons I don't
understand. In this case, it was the title attribute of a lone link
tag. That caused the line to be a unicode and then suddenly everything
died at the ''.join(out).

This fixes that by making sure everything is a utf-8 encoded string.